+++ src/lib-components/RichText.vue
@@ -6,7 +6,10 @@
 </template>
 
 <script>
+// UTILITY FUNCTIONS
 import stripCraftURLFromText from "@/mixins/stripCraftURLFromText"
+import accessibleExternalLinks from "@/mixins/accessibleExternalLinks"
+
 export default {
     name: "RichText",
     components: {},
@@ -16,10 +19,12 @@ export default {
             default: "",
         },
     },
-    mixins: [stripCraftURLFromText],
+    mixins: [stripCraftURLFromText, accessibleExternalLinks],
     computed: {
         parsedContent() {
-            return this.stripCraftURLFromText(this.richTextContent)
+            let content = this.stripCraftURLFromText(this.richTextContent)
+
+            return this.accessibleExternalLinks(content)
         },
     },
 }
@@ -135,6 +140,22 @@ export default {
             @include link-hover;
         }
     }
+
+    ::v-deep a[target="_blank"] {
+        position: relative;
+        margin-right: 25px;
+    }
+
+    ::v-deep a[target="_blank"]:after {
+        content: url("node_modules/ucla-library-design-tokens/assets/svgs/icon-external-link.svg");
+        display: inline-block;
+        margin-left: 0.2em;
+        position: absolute;
+        bottom: 7.3px;
+        width: 1em;
+        height: 1em;
+    }
+
     ::v-deep ul,
     ::v-deep ol {
         padding: 0 16px;