+++ src/mixins/stripCraftURLFromText.js
@@ -11,12 +11,18 @@ export default {
                 return text // don't try string method .replace
             }
             if (text.indexOf("https://test-craft.library.ucla.edu") !== -1)
-                return text.replace("https://test-craft.library.ucla.edu", "")
+                return text.replaceAll(
+                    "https://test-craft.library.ucla.edu",
+                    ""
+                )
             else if (
                 text.indexOf("https://stage-craft.library.ucla.edu") !== -1
             )
-                return text.replace("https://stage-craft.library.ucla.edu", "")
-            else return text.replace("https://craft.library.ucla.edu", "")
+                return text.replaceAll(
+                    "https://stage-craft.library.ucla.edu",
+                    ""
+                )
+            else return text.replaceAll("https://craft.library.ucla.edu", "")
         },
     },
 }