+++ src/stories/SectionTeaserList.stories.js
@@ -54,6 +54,39 @@ const mock = [
         sectionHandle: "event",
         text: "Accumsan tortor posuere ac ut consequat semper viverra libero.",
     },
+    {
+        image: API.image,
+        to: "/visit/foo/bar/",
+        category: "Ullamco",
+        title: "Fames ac turpis egestas sed tempus lorem ipsum",
+        startDate: "2021-09-03T08:00:00+00:00",
+        endDate: "2021-10-08T07:10:00+00:00",
+        text: "Ultricies leo integer malesuada nunc vel risus commodo viverra.",
+        sectionHandle: "event",
+    },
+]
+
+const mixedOngoingEvents = [
+    {
+        image: API.image,
+        to: "/visit/foo/bar/",
+        category: "Ullamco",
+        title: "Fames ac turpis egestas sed tempus lorem ipsum",
+        startDate: "2021-09-03T08:00:00+00:00",
+        endDate: "2021-10-08T07:10:00+00:00",
+        text: "Ultricies leo integer malesuada nunc vel risus commodo viverra.",
+        sectionHandle: "event",
+    },
+
+    {
+        image: API.image,
+        to: "/visit/foo/bar/",
+        category: "Ullamco",
+        title: "Fames ac turpis egestas sed tempus lorem ipsum",
+        text: "Ultricies leo integer malesuada nunc vel risus commodo viverra.",
+        sectionHandle: "event",
+        ongoing: true,
+    },
 ]
 
 // Variations of stories below
@@ -73,6 +106,18 @@ export const Default = () => ({
 export const Expandable = () => ({
     data() {
         return { items: mock }
+    },
+    components: { SectionTeaserList },
+    template: `
+      <section-teaser-list
+        :items="items"
+      />
+  `,
+})
+
+export const MixedOngoingEvents = () => ({
+    data() {
+        return { items: mixedOngoingEvents }
     },
     components: { SectionTeaserList },
     template: `